Fixed state mismatch in glossary window; fixed Japanese language in templates
The glossary window doesn't have a review state, but it does have a special glossary lookup state. (That could actually probably be folded in with the match state; and we could probably make use of a review state...)

Since there are no memory/glossary versions of the gloss state and review state, respectively, I folded up the subclasses for those as well.